+++ lib/config/constants/constants.dart
@@ -1,0 +1,21 @@
+import 'package:limetv/config/constants/assets.dart';
+
+List<String> settingsMenuName = [
+  'Основные',
+  'Оплата услуг',
+  'Мои утсройства',
+  'Защита от детей',
+  'Жалобы и предложения',
+  'Уведомление',
+  'Выход',
+];
+
+List<String> settingIconText = [
+  Assets.icons.settingsMain,
+  Assets.icons.paymentServices,
+  Assets.icons.myDevices,
+  Assets.icons.securityChilds,
+  Assets.icons.complaintsSuggestions,
+  Assets.icons.notifications,
+  Assets.icons.exit,
+];
