+++ lib/presentation/pages/settings_page/settings_page.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 import 'package:limetv/config/constants/app_colors.dart';
 import 'package:limetv/config/constants/app_text_styles.dart';
-import 'package:limetv/config/constants/assets.dart';
+import 'package:limetv/config/constants/constants.dart';
 import 'package:limetv/presentation/components/genre_label.dart';
 import 'package:limetv/presentation/components/web_appbar.dart';
+import 'package:limetv/presentation/pages/settings_page/widget/settings_menu_item.dart';
 
 class SettingsPage extends StatefulWidget {
   const SettingsPage({super.key});
@@ -17,32 +17,14 @@ class SettingsPage extends StatefulWidget {
 class _SettingsPageState extends State<SettingsPage> {
   int onTap = 0;
 
-  List<String> settingsMenuName = [
-    'Основные',
-    'Оплата услуг',
-    'Мои утсройства',
-    'Защита от детей',
-    'Жалобы и предложения',
-    'Уведомление',
-    'Выход',
-  ];
-
-  List<String> settingIconText = [
-    Assets.icons.settingsMain,
-    Assets.icons.paymentServices,
-    Assets.icons.myDevices,
-    Assets.icons.securityChilds,
-    Assets.icons.complaintsSuggestions,
-    Assets.icons.notifications,
-    Assets.icons.exit,
-  ];
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: AppColors.backgroundColor,
       appBar: const WebAppBar(),
       body: Row(
+        mainAxisAlignment: MainAxisAlignment.start,
+        crossAxisAlignment: CrossAxisAlignment.start,
         children: [
           Container(
             width: 450.w,
@@ -67,7 +49,7 @@ class _SettingsPageState extends State<SettingsPage> {
                         onTap: () {
                           setState(() {
                             onTap = index;
-                          });                                                    
+                          });
                         },
                         child: SettingsMenuItem(
                           text: settingsMenuName[index],
@@ -78,10 +60,19 @@ class _SettingsPageState extends State<SettingsPage> {
                       );
                     },
                   ),
-                ),                
+                ),
               ],
             ),
           ),
+          Container(
+            margin: EdgeInsets.only(top: 162.h, left: 72.w),
+            child: Text(
+              settingsMenuName[onTap],
+              style: AppTextStyles.body37w5.copyWith(
+                color: Colors.white,
+              ),
+            ),
+          )
         ],
       ),
     );
@@ -96,61 +87,4 @@ Color colorFunc(int index, int onTap) {
     return const Color(0xffF2994A);
   }
   return const Color(0xff828282);
-}
-
-
-
-class SettingsMenuItem extends StatelessWidget {
-  SettingsMenuItem({
-    required this.text,
-    required this.color,
-    required this.assetsIconText,
-    required this.onClick,
-    Key? key,
-  }) : super(key: key);
-
-  String text;
-  String assetsIconText;
-  Color color;
-  bool onClick;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      alignment: Alignment.centerLeft,
-      padding: EdgeInsets.symmetric(vertical: 16.h),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          onClick ? SvgPicture.asset(
-            Assets.icons.point,
-            color: color,
-          ) : SvgPicture.asset(
-            Assets.icons.point,
-            color: Colors.transparent,
-          ) ,
-          Row(
-            children: [
-              SizedBox(
-                width: 17.3.w,
-              ),
-              SvgPicture.asset(
-                assetsIconText,
-                color: color,
-              ),
-              SizedBox(
-                width: 17.3.w,
-              ),
-              Text(
-                text,
-                style: AppTextStyles.body18w5.copyWith(
-                  color: color,
-                ),
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
 }
